(PC-20633)[PRO] refactore: Add new onboarding user redirection test and refactor with typescript
